Implementación del DatePicker y Combo box de género en la Vista registro, además de implementación de "." automatico en los campos que ocupen decimales

+++ app/src/main/java/ni/edu/uca/systemfits/modelo/Medidas.kt
@@ -8,5 +8,5 @@ data class Medidas(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
     val musculo: String,
-    val tamaño: Int,
+    val tamaño: Double,
 )
